+++ R/MCMCpstr.R
@@ -10,7 +10,9 @@
 #'
 #' @param excl Character string (or vector of character strings) denoting parameters to exclude. Used in conjunction with \code{params} argument to select parameters of interest.
 #'
-#' @param ISB Ignore Square Brackets (ISB). Logical specifying whether square brackets should be ignored in the \code{params} and \code{excl} arguments. If \code{TRUE}, square brackets are ignored - input from \code{params} and \code{excl} are otherwise matched exactly. If \code{FALSE}, square brackets are not ignored - input from \code{params} and \code{excl} are matched using grep, which can take arguments in regular expression format. This allows partial names to be used when specifying parameters of interest.
+#' @param ISB Ignore Square Brackets (ISB). Logical specifying whether square brackets should be ignored in the \code{params} and \code{excl} arguments. If \code{TRUE}, square brackets are ignored. If \code{FALSE}, square brackets are not ignored.  This allows partial names to be used when specifying parameters of interest. Use \code{exact} argument to specify whether input from \code{params} and \code{excl} arguments should be matched exactly.
+#'
+#' @param exact Logical specifying whether input from \code{params} and \code{excl} arguments should be matched exactly (after ignoring square brackets if \code{ISB = FALSE}). #' If \code{TRUE}, input from \code{params} and \code{excl} are matched exactly (after taking \code{ISB} argument into account). If \code{FALSE}, input from \code{params} and \code{excl} are matched using regular expression format (after taking \code{ISB} argument into account).
 #'
 #' @param func Function to be performed on MCMC output. When output of specified function is greater than length 1, an extra dimension is added. For instance, output of length 3 for a parameter with dimensions 2x2 results in a 2x2x3 output. Functions that produce output with dimensionality greater than 1 are not permitted. \code{func} is ignored when \code{type = 'chains'}.
 #'
@@ -31,15 +33,16 @@ MCMCpstr <- function(object,
                    params = 'all',
                    excl = NULL,
                    ISB = TRUE,
+                   exact = TRUE,
                    func = mean,
                    type = 'summary')
 {
   #SORTING BLOCK
   if (methods::is(object, 'matrix'))
   {
-    object2 <- MCMCchains(object, params, excl, ISB, mcmc.list = FALSE)
+    object2 <- MCMCchains(object, params, excl, ISB, exact = exact, mcmc.list = FALSE)
   } else {
-    object2 <- MCMCchains(object, params, excl, ISB, mcmc.list = TRUE)
+    object2 <- MCMCchains(object, params, excl, ISB, exact = exact, mcmc.list = TRUE)
   }
 
   if (coda::is.mcmc.list(object2) == TRUE)