+++ NEWS.md
@@ -3,6 +3,7 @@ NEWS
 
 0.15.0:
 
+- All functions now take `exact` as an argument to parse desired parameter. `ISB = TRUE` no longer uses regex matching (`exact` should be used to specify this).
 - Add support for objects output from the `nimble` package
 -`MCMCpstr` fix bug where the incorrect number of dimensions would be returned when subsetting a param with >1 dim